1. fixed script_train.sh 2. add setup.sh, eval_pipeline.sh, inference_pipeline.sh 3. add scale parameter in render_mp4 function 4. add some comments 5. add dataset generation scripts 6. fixed sandbox_fid.py and update get_motion_scalers.sh 7. fixed: always generating a empty folder in inference 8. add eval_beat_metrics.py and feat2motion.py

+++ training/options/base_options.py
@@ -150,11 +150,13 @@ class BaseOptions:
         utils.mkdirs(expr_dir)
         file_name = os.path.join(expr_dir, 'opt.txt')
         file_name_json = os.path.join(expr_dir, 'opt.json')
-        with open(file_name, 'wt') as opt_file:
-            opt_file.write(message)
-            opt_file.write('\n')
-        with open(file_name_json, 'wt') as opt_file:
-            opt_file.write(json.dumps(vars(opt)))
+        if not os.path.exists(file_name):
+            with open(file_name, 'wt') as opt_file:
+                opt_file.write(message)
+                opt_file.write('\n')
+        if not os.path.exists(file_name_json):
+            with open(file_name_json, 'wt') as opt_file:
+                opt_file.write(json.dumps(vars(opt)))
 
     def parse(self, parse_args=None):
